+++ Day1/task2b.c
@@ -3,11 +3,11 @@
 
 int main()
 {
-   float r,l,h;
+   double r,l,h;
    fprintf(stdout,"Enter R ,L and H: ");
-   scanf("%f %f %f",&r,&l,&h);
-   float area=acosf((r-h)/h)*r*r-(r-h)*sqrtf(2*r*h-h*h);
-   float volume=area*l;
-   fprintf(stdout,"Water in tank is %f liters\n",volume);
+   scanf("%lf %lf %lf",&r,&l,&h);
+   double area=acos((r-h)/h)*r*r-(r-h)*sqrt(2*r*h-h*h);
+   double volume=area*l;
+   fprintf(stdout,"Water in tank is %lf liters\n",volume);
    return 0;
 }
